FS-8783: [libsrtp] Fix alignment issue
Backport upstream alignment fix to correct bus error on platforms
that require strict memory alignment such as SPARC

FS-8783 #resolve

From upstream:
commit 4d8430a504137509f23b5a19f8a06b6df0f651cc
Author: Jaap Keuter <jaap.keuter@xs4all.nl>
Date:   Fri Nov 7 00:13:10 2014 +0100

While setting the IV for AES ICM the nonce is simply typecast from
a void * to a v128_t *. This breaches alignment requirements for
v128_t objects on platforms that require it.
Instead make a copy of the nonce to assure proper alignment.