FS-10617: [mod_sofia] Nightmare transfer: expose remote server's channel UUIDs as variables.
Currently there is no easy way for ESL and dial plan users to easily correlate
the channel UUID of the call legs involved in a nightmare transfer. This patch
adds two new channel variables, which are set to the remote server call leg
UUIDs (transfer_refer_from, transfer_refer_for).

The UUIDs are passed from the remote server in custom headers (X-FS-Refer-From
and X-FS-Refer-For).